backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.3
python-dotenv>=1.0.0
openai>=1.10.0
google-generativeai>=0.8.0
anthropic>=0.40.0
requests>=2.32.0
